Flatten the class hierarchy
